+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## v.1.20.x
+- Changed AdaptivePolicyGroup CreatedAt and UpdatedAt to be DateTime
+- Changed AdaptivePolicyPolicies to correct nullability. Changed timestamps to DateTime
+- Updated RoutingOspf to add RoutingOspfV3 settings
+
 ## v1.20.42
 - v1.15 changes
 - Added new properties to several models.